修复Subsonic音乐艺术家歌曲数据获取问题

+++ api/src/main/java/cn/xybbz/api/client/subsonic/data/AlbumID3.kt
@@ -48,7 +48,7 @@ data class AlbumID3  (
     /**
      * 持续时间
      */
-    val duration: Long,
+    val duration: Float,
     /**
      * 歌曲数量
      */
